requirements.txt
streamlit
requests
mysql-connector-python
pandas
geopy
matplotlib
reportlab
pytest